Created Level.stepCompleted(), Level.cs comments added
The function stepCompleted() has been added to the Level class. This funciton will return bool True when a step has been completed on the current frame and False otherwise. This function is utilized in ProgressionTracker.cs. For the Progression Tracking 'level' this function is very simple because, for testing purposes, steps are completed by pressing the 'X' button. For future levels however this function will be much more complex, likely relying on the scene's Level class for its stepCount value which will create a cross dependency. Also added general commenting to ProgressionTracker.cs for clarity.