Integrated Progression Tracking 'level'
The Progression Tracking 'level' has been added to the main menu through an appropriate addition to SceneChange.cs, the addition of the Progression Tracking scene to the build settings, and the addition of a "Progression Tracking" button in the Menu scene. The contents of the Progression Tracking scene have been moved to better match up with the starting location set by the Menu scene. I've also updated the game title as displayed in the menu scene.