Flashcard scene changes, deck changing is now immediate with an indication sound, added handling for displaying long file names
I've added the whiteboard previously used in the progression tracking scene to hold the scene buttons as well as the instructional text which has been updated and will no longer display on the omnicard itself. Its removal from the omnicard will slightly simplify the logic for the 'learned' card system which is still a bit buggy atm. A deck change will now immediatley display on the omnicard with an accompanying indication sound. Long filenames will now display truncated as to avoid text overflowing.