Debugged 'learned card' flashcard functionality
The system for hiding learned cards from the current deck has been debugged. The system appears to be working as intended now but more rigourous testing is needed to confirm this. Bugs related to this system may have been causing the bugs in the flashcard randomization functionality but again, further testing is needed on that front.